requirements.txt
langchain-community>=0.4.1
langchain-core>=1.2.9
langchain-text-splitters>=1.1.0
langgraph>=1.0.8
dashscope>=1.25.11
faiss-cpu>=1.13.2
python-dotenv>=1.2.1
httpx>=0.28.1
python-docx>=1.2.0
openpyxl>=3.1.5
python-pptx>=1.0.2
